+++ e2e/__tests__/coverageTransformInstrumented.test.ts
@@ -21,7 +21,7 @@ it('code coverage for transform instrumented code', () => {
   run('yarn', dir);
   const result = runJest(dir, ['--coverage', '--no-cache']);
 
-  expect(result.status).toBe(0);
+  expect(result.exitCode).toBe(0);
 
   const coverageMapFile = path.join(coverageDir, 'coverage-final.json');
   const coverageMap = JSON.parse(readFileSync(coverageMapFile, 'utf-8'));