+++ e2e/__tests__/emptyDescribeWithHooks.test.ts
@@ -16,24 +16,24 @@ skipSuiteOnJasmine();
 
 test('hook in empty describe', () => {
   const result = runJest(dir, ['hookInEmptyDescribe.test.js']);
-  expect(result.status).toBe(1);
+  expect(result.exitCode).toBe(1);
   expect(extractSummary(result.stderr)).toMatchSnapshot();
 });
 
 test('hook in describe with skipped test', () => {
   const result = runJest(dir, ['hookInDescribeWithSkippedTest.test.js']);
-  expect(result.status).toBe(0);
+  expect(result.exitCode).toBe(0);
   expect(extractSummary(result.stderr)).toMatchSnapshot();
 });
 
 test('hook in empty nested describe', () => {
   const result = runJest(dir, ['hookInEmptyNestedDescribe.test.js']);
-  expect(result.status).toBe(1);
+  expect(result.exitCode).toBe(1);
   expect(extractSummary(result.stderr)).toMatchSnapshot();
 });
 
 test('multiple hooks in empty describe', () => {
   const result = runJest(dir, ['multipleHooksInEmptyDescribe.test.js']);
-  expect(result.status).toBe(1);
+  expect(result.exitCode).toBe(1);
   expect(extractSummary(result.stderr)).toMatchSnapshot();
 });