+++ e2e/__tests__/testTodo.test.ts
@@ -13,35 +13,35 @@ const dir = path.resolve(__dirname, '../test-todo');
 
 test('works with all statuses', () => {
   const result = runJest(dir, ['statuses.test.js']);
-  expect(result.status).toBe(1);
+  expect(result.exitCode).toBe(1);
   const {rest} = extractSummary(result.stderr);
   expect(wrap(rest)).toMatchSnapshot();
 });
 
 test('shows error messages when called with no arguments', () => {
   const result = runJest(dir, ['todoNoArgs.test.js']);
-  expect(result.status).toBe(1);
+  expect(result.exitCode).toBe(1);
   const {rest} = extractSummary(result.stderr);
   expect(wrap(rest)).toMatchSnapshot();
 });
 
 test('shows error messages when called with multiple arguments', () => {
   const result = runJest(dir, ['todoMultipleArgs.test.js']);
-  expect(result.status).toBe(1);
+  expect(result.exitCode).toBe(1);
   const {rest} = extractSummary(result.stderr);
   expect(wrap(rest)).toMatchSnapshot();
 });
 
 test('shows error messages when called with invalid argument', () => {
   const result = runJest(dir, ['todoNonString.test.js']);
-  expect(result.status).toBe(1);
+  expect(result.exitCode).toBe(1);
   const {rest} = extractSummary(result.stderr);
   expect(wrap(rest)).toMatchSnapshot();
 });
 
 test('shows todo messages when in verbose mode', () => {
   const result = runJest(dir, ['verbose.test.js', '--verbose']);
-  expect(result.status).toBe(0);
+  expect(result.exitCode).toBe(0);
   const {rest} = extractSummary(result.stderr);
   expect(wrap(rest)).toMatchSnapshot();
 });