+++ examples/timer/__tests__/infinite_timer_game.test.js
@@ -19,7 +19,7 @@ it('schedules a 10-second timer after 1 second', () => {
   // (but not any new timers that get created during that process)
   jest.runOnlyPendingTimers();
 
-  // At this point, our 1-second timer should have fired it's callback
+  // At this point, our 1-second timer should have fired its callback
   expect(callback).toBeCalled();
 
   // And it should have created a new timer to start the game over in